update: added functionality to open into a new tab if presses with alt / ctrl key
