--- frontend/src/app/data/JobListinTemp.js
@@ -1,178 +1,0 @@
-const jobListings = [
-    {
-        title: 'Help wdwadwadawdawkjdhakwjdhwajhdkwajhdkjwahw',
-        location: 'London, UK, WV10 9QL',
-        description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-        tags: [
-            { name: '£30hr, 7hr / week' },
-            { name: 'Starts: 12/05/2024' },
-            { name: 'Starts: 12/05/2024' },
-            { name: 'Starts: 12/05/2024' },
-            { name: 'Starts: 12/05/2024' },
-        ],
-        person: 'John Doe',
-        postedOn: '21st August 2021'
-    },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-    // {
-    //     title: 'Help with childcare',
-    //     location: 'London, UK, WV10 9QL',
-    //     description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, dignissimos. Rem vitae quos nulla aliquam molestiae ipsam, doloribus repellendus omnis quas officiis provident possimus eligendi culpa veritatis voluptas unde fugiat.',
-    //     tags: [
-    //         { name: '£30hr, 7hr / week' }
-    //     ],
-    //     person: 'John Doe',
-    //     postedOn: '21st August 2021'
-    // },
-]
-
-export default jobListings;
